Turned health into a property and made the health bar a slider object.
Made it so now every time you want to change health all you have to do is add or subtract from it and it will update the health bar and check if you die. Also added a "Helpers" class that hopefully will have a random array of functions that can be used in many different places and will act as the static library of helpful things.