NOT WORKING
working on adding a system for automatic animation end functions to be called. The problem currently is that delegates are not passed as data types and not references or some bull crap like that so I can't alter them. Solution is going to be make them properties and whenever they get changed just update the list and regular delegate at the same time....