Cleaned up the UI Object and made it a single working prefab linked into the new UI manager
The new UI manager is a script on the high level UI object that works by enabling and disabling associated objects to each Ui page and managing which get shown and when. Additionally, there are unity events that can be trigged on enable or disable to allow for more complex ui pages like the inventory where the script is located under the player and handled independently of much of the other ui objects. but not anymore!